fix: Locations remove error handling

+++ crm_backend/src/main/scala/locations/application/remove_location/RemoveLocationUseCase.scala
@@ -7,5 +7,8 @@ import products.domain.repository.ProductRepository
 class RemoveLocationUseCase()(using locationRepository: LocationRepository, productRepository: ProductRepository) extends BaseUseCase[RequestRemoveLocation, ResponseRemoveLocation]:
 
   override def execute(request: RequestRemoveLocation): Option[ResponseRemoveLocation] =
+    val amountOfProductsInLocation = productRepository.getProductsWithLocation(request.id).size
+    if(amountOfProductsInLocation > 0)
+      return Some(ResponseRemoveLocation(data = "Can't remove location because it have products"))
     locationRepository.removeLocation(request.id) 
     Some(ResponseRemoveLocation(data= "Removed"))
